Specified some tests...to complete later
- expect this to be messy until I figure out how to write them better
